+++ doc/history.md
@@ -5,5 +5,6 @@
 | 0.1     | 03.06.2026 | Initial version |
 | 0.2     | 07.06.2026 | Dynamic memory cleaning added.<br>Memory usage dropped from 2-20GB to 0.2-2GB depends on config.<br>Also fractal.h → basics.h |
 | 0.3     | 09.06.2026 | Title change to: Cosmic Fractal<br>Pause Timing corrected |
-| 0.4     | 15.06.2026 | Full Screen Mode added (-f).<br>Also Resize window.<br> Release v.0.4.0. |
+| 0.4     | 15.06.2026 | Full Screen Mode added (-f).<br>Also Resize window.<br> Release v0.4.0. |
 | 0.4.1   | 26.06.2026 | `--fps` option added |
+| 0.5.0   | 28.06.2026 | Auto Scale more smooth animation.<br> Release v0.5.0 |